refactor(cli): refactor create command style
