+++ packages/varlet-cli/package.json
@@ -55,7 +55,6 @@
     "execa": "^5.0.0",
     "fs-extra": "^9.0.1",
     "glob": "^7.2.0",
-    "gradient-string": "^2.0.1",
     "hash-sum": "^2.0.0",
     "inquirer": "^8.1.2",
     "jest": "^26.6.3",
@@ -63,7 +62,6 @@
     "lodash": "^4.17.21",
     "lodash-es": "^4.17.21",
     "ora": "^5.4.0",
-    "prettier": "^2.7.1",
     "prompts": "^2.4.2",
     "semver": "^7.3.5",
     "slash": "^3.0.0",
