--- packages/varlet-cli/src/shared/options.ts
@@ -1,12 +1,0 @@
-interface CreateOptions {
-  name: string
-  locale?: boolean
-  projectName?: string
-  style?: string
-  namespace?: string
-}
-const createOptions: CreateOptions = {
-  name: 'componentName',
-}
-
-export { createOptions }
