+++ packages/varlet-cli/template/create/jsx.ejs
@@ -1,4 +1,4 @@
-import './<%- name %>.less'
+import './<%- projectName %>.less'
 import { props } from './props'
 
 export default ({
