+++ packages/varlet-cli/template/create/tsx.ejs
@@ -1,6 +1,6 @@
 import { defineComponent } from 'vue'
 import { props } from './props'
-import './<%- name %>.less'
+import './<%- projectName %>.less'
 
 export default defineComponent({
   setup(props, { slots }) {
