+++ packages/varlet-cli/template/create/vue.ejs
@@ -4,7 +4,7 @@
 
 <script lang="ts">
   import { defineComponent } from 'vue'
-  import './<%- name %>.less'
+  import './<%- projectName %>.less'
 
   export default defineComponent({
     name: '<%- namespace + name %>',
